fix: 修正资源系统逻辑错误
- AssetHandle::IsValid(): || 改为仅检查 Index!=0（Slot 0 永久保留）
- SlotMap::Size(): 追踪活跃元素数而非历史最大值，新增 Capacity()
- BIT 宏: 添加 #ifndef 防护避免双重定义
- s_GLFWWindowCount: uint8_t 改为 uint32_t 防溢出
- EditorAssetDescriptor: 未注册类型回退时输出 warning 日志
- 同步更新 SlotMap 单元测试

+++ Engine/src/Core/Window.cpp
@@ -26,7 +26,7 @@
 namespace Engine
 {
 
-    static uint8_t s_GLFWWindowCount = 0;
+    static uint32_t s_GLFWWindowCount = 0;
 
 #ifdef __linux__
     static int X11ErrorHandler(Display*, XErrorEvent* event)
